isEnvironmentExists util; tests for isEnvironmentExists; typo fix in env name; readme update
