+++ packages/prepare-environment/.gitignore
@@ -2,3 +2,4 @@ node_modules
 build
 environments/*
 .env.production
+.idea
