+++ packages/prepare-environment/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@37bytes/prepare-environment",
-  "version": "1.0.4",
+  "version": "2.0.0",
   "description": "Utility to prepare the environment for a React application. Useful for CI.",
   "repository": "https://github.com/37bytes/public-packages/tree/master/packages/prepare-environment",
   "author": "37bytes",
